Move shader buffered sraw node into single class. Because both karaoke and lyric sprite text will use same draw node.
