+++ osu.Framework.Font/Graphics/Sprites/LyricSpriteText_DrawNode.cs
@@ -3,10 +3,7 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using osu.Framework.Graphics.Colour;
-using osu.Framework.Graphics.OpenGL;
-using osu.Framework.Graphics.OpenGL.Buffers;
 using osu.Framework.Graphics.OpenGL.Vertices;
 using osu.Framework.Graphics.Primitives;
 using osu.Framework.Graphics.Shaders;
@@ -43,18 +40,17 @@ namespace osu.Framework.Graphics.Sprites
         }
 
         // todo: should have a better way to let user able to customize formats?
-        protected override DrawNode CreateDrawNode() => new LyricSpriteTextShaderEffectDrawNode(this, new LyricSpriteTextShaderEffectDrawNodeSharedData(null, false));
+        protected override DrawNode CreateDrawNode()
+            => new LyricSpriteTextShaderEffectDrawNode(this, new LyricSpriteTextShaderEffectDrawNodeSharedData(null, false));
 
         /// <summary>
         /// <see cref="BufferedDrawNode"/> to apply <see cref="IShader"/>.
         /// </summary>
-        protected class LyricSpriteTextShaderEffectDrawNode : BufferedDrawNode
+        protected class LyricSpriteTextShaderEffectDrawNode : MultiShaderBufferedDrawNode
         {
             protected new LyricSpriteText Source => (LyricSpriteText)base.Source;
 
             private long updateVersion;
-
-            private IReadOnlyList<IShader> shaders;
 
             public LyricSpriteTextShaderEffectDrawNode(LyricSpriteText source, LyricSpriteTextShaderEffectDrawNodeSharedData sharedData)
                 : base(source, new LyricSpriteTextDrawNode(source), sharedData)
@@ -66,53 +62,9 @@ namespace osu.Framework.Graphics.Sprites
                 base.ApplyState();
 
                 updateVersion = Source.updateVersion;
-                shaders = Source.Shaders;
             }
 
             protected override long GetDrawVersion() => updateVersion;
-
-            protected override void PopulateContents()
-            {
-                base.PopulateContents();
-                drawFrameBuffer();
-            }
-
-            protected override void DrawContents()
-            {
-                DrawFrameBuffer(SharedData.CurrentEffectBuffer, DrawRectangle, Color4.White);
-            }
-
-            private void drawFrameBuffer()
-            {
-                if (!shaders.Any())
-                    return;
-
-                FrameBuffer current = SharedData.CurrentEffectBuffer;
-                FrameBuffer target = SharedData.GetNextEffectBuffer();
-
-                GLWrapper.SetBlend(BlendingParameters.None);
-
-                foreach (var shader in shaders)
-                {
-                    var isFirst = shaders.ToList().IndexOf(shader) == 0;
-                    var source = isFirst ? current : target;
-
-                    using (BindFrameBuffer(target))
-                    {
-                        UpdateUniforms(shader, source);
-
-                        shader.Bind();
-                        DrawFrameBuffer(source, new RectangleF(0, 0, source.Texture.Width, source.Texture.Height), ColourInfo.SingleColour(Color4.White));
-                        shader.Unbind();
-                    }
-                }
-            }
-
-            protected virtual void UpdateUniforms(IShader targetShader, FrameBuffer current)
-            {
-                if (targetShader is ICustomizedShader customizedShader)
-                    customizedShader.ApplyValue(current);
-            }
         }
 
         public class LyricSpriteTextShaderEffectDrawNodeSharedData : BufferedDrawNodeSharedData
